Moved image tool result components to a dedicated vercel chat feature dir.
