+++ components/VercelChat/tools/image/ImageSkeleton.tsx
